Move generic templated sorts to header files
