+++ numpy/_core/src/common/npy_sort.h.src
@@ -26,7 +26,7 @@ extern "C" {
 
 /*
  *****************************************************************************
- **                        NEW SORT METHOD REGISTRATIONS                    **
+ **                       NEW SORT METHOD REGISTRATIONS                     **
  *****************************************************************************
  */
